+++ HeatingControl/Application/Loops/Processing/HysteresisProcessor.cs
@@ -2,12 +2,12 @@
 {
     public interface IHysteresisProcessor
     {
-        bool Process(float currentTemperature, bool currentState, float setPoint, float hysteresis);
+        bool Process(double currentTemperature, bool currentState, float setPoint, float hysteresis);
     }
 
     public class HysteresisProcessor : IHysteresisProcessor
     {
-        public bool Process(float currentTemperature, bool currentState, float setPoint, float hysteresis)
+        public bool Process(double currentTemperature, bool currentState, float setPoint, float hysteresis)
         {
             // TODO: can handle both heating and cooling approach
